requirements.txt
annotated-types==0.5.0
anyio==3.7.1
APScheduler==3.10.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
fastapi==0.103.1
h11==0.14.0
idna==3.4
loguru==0.7.1
mojito2==0.1.6
pyaes==1.6.1
pyasn1==0.5.0
pycryptodome==3.18.0
pydantic==2.3.0
pydantic_core==2.6.3
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
Telethon==1.29.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
uvicorn==0.23.2
websockets==11.0.3
win32-setctime==1.1.0
